Add functions for unit testing processTrigger
processTriggerSpec takes a Interface instead of a concrete Client.
This allows us to use a fake Client for unit tests.

This commit adds a GetFakeTriggersClient method which sets up a fake
client and seeds it with any resources such as TriggerBindings that
we might need for testing.

Also, added an example of passing TriggerBinding, TriggerTemplate
needed for unit tests using test.Resources.